refactor: some coding parameters

+++ src/Candidate.groovy
@@ -2,7 +2,7 @@ class Candidate extends AbstractPerson {
     String personalID // CPF
     int age
 
-    Candidate(String name, String email, String personalID, short age, String state, String zipCode, String description, List<String> skills) {
+    Candidate(String name, String email, String personalID, int age, String state, String zipCode, String description, List<String> skills) {
         this.name = name
         this.email = email
         this.personalID = personalID
